chore: reorganize documentation and clean repository
- Reorganize docs into structured subdirectories (guides, reference, architecture)
- Consolidate temporary docs into .archive/ (week-logs, phase-logs, session-logs)
- Remove 27 redundant documents (checklists, status, summaries)
- Update .gitignore: add __pycache__, exported_data/, teacher_data/, results/
- Root markdown: 42 → 7 files (83% reduction)
- Archive: 212 → 185 files (14% reduction)
- All tests passing: 790/790 library tests, Python linting clean

+++ tests/phase_6_integration.rs
@@ -452,5 +452,5 @@ fn test_recovery_after_circuit_breaker_reset() {
     // Reset
     cb.reset();
     assert_eq!(cb.state(), CircuitState::Closed);
-    assert!((cb.failure_ratio() as f64).abs() < f64::EPSILON);
+    assert!(cb.failure_ratio().abs() < f64::EPSILON);
 }
